fix: resolve all residual CI issues
- Fix test_save_snapshot_writes_json for cross-platform path compat
- Add docs/architecture/_template.md to mkdocs.yml nav
- Create .github/workflows/deploy-docs.yml for GitHub Pages
- Fix all 45 ruff ARG/SIM violations and re-enable rule sets
- Fix all 77 mypy type errors across 22 files and 24 modules
- Remove all mypy ignore_errors overrides

+++ src/osiris_toolkit/analysis/density.py
@@ -59,7 +59,7 @@ class DensityAnalyzer(DiagnosticAnalyzer):
             n = grid.data.shape[axis]
             coord = np.linspace(grid.axes[axis].min, grid.axes[axis].max, n)
         else:
-            coord = np.arange(len(profile))
+            coord = np.arange(len(profile), dtype=np.float64)
 
         return DensityProfileResult(
             species=species,
